Add notes regarding ettercap and xinetd service
Ettercap doesn't require IP Forwarding to be set. From the ettercap man page: "Every time ettercap starts, it disables ip forwarding in the kernel and begins to forward packets itself".
Ettercap's own forwarding of packets can fail, especially in the case of Telnet data packets, so running it with the -o flag delegates the forwarding to the kernel which seems to be more reliable.
Also added a note regarding the location of the telnetd executable possibly being different (/usr/sbin/telnetd in my case).